# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventSpot API Documentation",
        "description": "API Documentation for EventSpot, a platform to discover and manage events.",
        "contact": {
            "email": "admin@eventspot.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "EventSpot API Server"
        }
    ],
    "paths": {
...
        "/api/events/{event_id}/favorite": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Check if an event is favorited by the authenticated user",
                "description": "Returns the favorite status of a specific event for the logged-in user.",
                "operationId": "9c251b56b52513e658eb6f3a761b6e0e",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID of the event to check favorite status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "This event is favorited."
                                                },
                                                "etat": {
                                                    "type": "integer",
                                                    "enum": [
                                                        0,
                                                        1
                                                    ],
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erreur": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Toggle favorite status for an event",
                "description": "Add or remove an event from the authenticated user's favorites.",
                "operationId": "__invoke",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID of the event to toggle favorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite status toggled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Event added to favorites."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Event removed from favorites."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erreur": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erreur": {
                                            "type": "string",
                                            "example": "Event not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/events/{event_id}/subscribers": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get active subscribers of an event",
                "description": "Retrieve the list of users subscribed (with active subscription) to a specific event. Only the event organizer can access this endpoint.",
                "operationId": "80036617dc8b3dbdfa57fbdf4cbda21a",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID of the event to get subscribers for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscribers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "event": {
                                            "type": "string",
                                            "example": "Event Title"
                                        },
                                        "total_abonnes": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "abonnes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only the event organizer can view subscribers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied. Only the organizer can view subscribers."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error retrieving subscribers: error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/events/{event_id}/subscribe": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Check subscription status for an event",
                "description": "Returns whether the authenticated user's subscription to a specific event is active or not.",
                "operationId": "05b0b2a5f4b2014e2d1a4af845aa842c",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID of the event to check subscription status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Your subscription to this event is active."
                                                },
                                                "etat": {
                                                    "type": "integer",
                                                    "enum": [
                                                        0,
                                                        1
                                                    ],
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erreur": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Toggle subscription status for an event",
                "description": "Activate, deactivate, or create a subscription for the authenticated user on a specific event.",
                "operationId": "931d1b1f5329b2fe60d59ebf1ff7f93d",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID of the event to toggle subscription",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription activated."
                                        },
                                        "etat": {
                                            "type": "integer",
                                            "enum": [
                                                0,
                                                1
                                            ],
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erreur": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InteretsResource": {
                "title": "InteretsResource",
                "description": "Représentation d'un intérêt",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "nom": {
                        "type": "string",
                        "example": "Musique"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "UserResource",
                "description": "Représentation d'un utilisateur",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-09T18:00:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "example": "09 July 2025"
                    },
                    "interets": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InteretsResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}